+++ app-agent/packages/integrations/vue/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @gakwaya/integrations-vue
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @gakwaya/entities@0.1.2
+  - @gakwaya/app-agent@0.1.3
+  - @gakwaya/integrations-shared@0.1.3
+  - @gakwaya/ui@0.1.2
+
 ## 0.1.2
 
 ### Patch Changes