+++ app-agent/packages/state-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gakwaya/state-manager
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @gakwaya/entities@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes